pyproject.toml
[tool.poetry]
name = "RecSys"
version = "0.1.0"
description = "MLOps project for building recommendations."
authors = ["alexwerben <crystall.werben@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
flake8 = "^7.1.1"
pytest = "^8.3.3"
wheel = "^0.44.0"
setuptools = "^75.1.0"
twine = "^5.1.1"
omegaconf = "^2.3.0"
boto3 = "^1.35.36"
scikit-learn = "^1.5.2"
pandas = "^2.2.3"
hydra-core = "^1.3.2"
rectools = "^0.8.0"
python-dotenv = "^1.0.1"
mlflow = "^2.16.2"
dvc = "^3.55.2"
dvc-s3 = "^3.2.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
